backend/requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
PyMySQL
python-dotenv
Werkzeug
PyJWT
